pyproject.toml
[project]
name = "yourbench"
version = "0.7.0"
authors = [
    { name = "Sumuk Shashidhar", email = "sumuks2@illinois.edu" },
    { name = "Alina Lozovskaia", email = "alina.lozovskaia@huggingface.co" },
    { name = "Clémentine Fourrier", email = "clementine@huggingface.co" },
    { name = "Nathan Habib", email = "nathan.habib@huggingface.co" },
]
description = "Dynamic Evaluation Set Generation with Large Language Models"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "asyncio>=3.4.3",
    "click>=8.1.7",
    "datasets>=3.3.0",
    "hf-transfer>=0.1.9",
    "huggingface-hub[inference,hf_xet]>=0.34.3",
    "hydra-core>=1.3.2",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "markitdown[all]>=0.0.2",
    "numpy>=1.24.0",
    "omegaconf>=2.3.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.0",
    "randomname>=0.2.1",
    "rich>=13.7.0",
    "ruff>=0.11.1",
    "thefuzz>=0.22.1",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
    "typer>=0.15.2",
]

[project.scripts]
yourbench = "yourbench.main:main"

[project.optional-dependencies]
llm = ["markitdown[all]>=0.0.2"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["yourbench*"]

[tool.setuptools.package-data]
yourbench = ["prompts/**/*.md", "utils/yourbench_card_template.md"]

[tool.ruff]
line-length = 119
exclude = ["**/*.ipynb"]

lint.ignore = ["E501", "C901", "F841"]
lint.select = ["C", "E", "F", "I", "W"]
lint.fixable = ["A", "B", "C", "D", "E", "F", "I", "W"]
preview = true

[tool.ruff.lint.isort]
length-sort = true
lines-after-imports = 2
no-lines-before = ["standard-library", "local-folder"]
known-local-folder = ["yourbench"]
known-first-party = ["datasets", "huggingface_hub", "transformers", "torch"]
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "pytest>=8.4.1",
    "ruff>=0.12.3",
    "twine>=6.1.0",
]
